Feat: CustomJSONDecoder 타입을 JSONDecodable로 추상화
- 클래스 타입을 추상화하여 JSONDecodable 프로토콜과 초기구현으로 변경
- Expo1900Tests 테스트 클래스에서 초기구현된 JSONDecodable 프로토콜 채택
- 프로토콜 채택에 따라 디코딩 방식 변경됨에 따라 테스트 케이스 로직 변경
